refactor:  improve code quality
