--- src/helpers/uuid-validater.ts
@@ -1,6 +1,0 @@
-export default function isValidUUID(uuid: string | undefined): boolean {
-  if (typeof uuid !== 'string') return false;
-  const reg =
-    /^[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$/i;
-  return reg.test(uuid);
-}
