+++ src/helpers/uuid-validator.ts
@@ -1,0 +1,8 @@
+export default function isValidUUID(uuid: string | undefined): boolean {
+  if (typeof uuid !== 'string' || uuid.length !== 36) {
+    return false;
+  }
+
+  const reg = /^[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$/i;
+  return reg.test(uuid);
+}
